optimize k8s nodeIp and k8s-cluster-id config
**Phenomenon and reproduction steps**

**Root cause and solution**

1. Get the Mac address from IP encapsulated into a function
2. Enable registration without configuring k8s-cluster-id

**Impactions**

**Test method**

**Affected branch(es)**

* main

**Checklist**

- [ ] Dependencies update required
- [ ] Common bug (similar problem in other repo)
