+++ agent/src/trident.rs
@@ -10,7 +10,7 @@ use std::sync::{
 use std::thread::{self, JoinHandle};
 use std::time::Duration;
 
-use anyhow::{anyhow, Context, Result};
+use anyhow::{Context, Result};
 use flexi_logger::{
     colored_opt_format, Age, Cleanup, Criterion, Duplicate, FileSpec, Logger, LoggerHandle, Naming,
 };
@@ -25,8 +25,7 @@ use crate::pcap::WorkerManager;
 use crate::utils::cgroups::Cgroups;
 use crate::utils::environment::{free_memory_check, get_k8s_local_node_ip};
 use crate::utils::guard::Guard;
-use crate::utils::net::addr_list;
-use crate::utils::net::link_list;
+use crate::utils::net::get_mac_by_ip;
 use crate::{
     collector::Collector,
     collector::{
@@ -161,35 +160,7 @@ impl Trident {
                     "use K8S_NODE_IP_FOR_METAFLOW env ip as destination_ip({})",
                     ip
                 );
-                let links = link_list()?;
-                let addrs = addr_list()?;
-                let if_idx = addrs
-                    .iter()
-                    .find_map(|a| {
-                        if a.ip_addr == ip {
-                            Some(a.if_index)
-                        } else {
-                            None
-                        }
-                    })
-                    .ok_or(anyhow!(format!(
-                        "can't find interface index by node ip {}",
-                        ip
-                    )))?;
-
-                let ctrl_mac = links
-                    .iter()
-                    .find_map(|l| {
-                        if l.if_index == if_idx {
-                            Some(l.mac_addr)
-                        } else {
-                            None
-                        }
-                    })
-                    .ok_or(anyhow!(format!(
-                        "can't find ctrl_mac address by node ip {}",
-                        ip
-                    )))?;
+                let ctrl_mac = get_mac_by_ip(ip)?;
                 (ip, ctrl_mac)
             }
             None => get_route_src_ip_and_mac(&config.controller_ips[0].parse()?)
